# whitebalance: optimize, remove SSE codepath