Optimized both plain and SSE code paths (there was a bug in Xtrans SSE that prevented full use of SSE, and mm_store is faster than mm_stream in Bayer SSE).  Found that plain C code was still as fast as or slightly faster than the SSE code even after optimizing the latter, so remove the SSE codepath for Bayer and Xtrans.  Keep SSE code path for non-mosaiced images for now.

Using two threads, Bayer with test 0000's sidecar goes from 6.9ms to 6.2ms, and Xtrans with mire1-xtrans.raf and test 0000's sidecar goes from 33ms to 22ms.  The gap narrows quickly at higher thread counts since the memory bus saturates.