@johnny-bit 2021-07-17
Great! Now the only CI failure is noise from mac. Can be ignored since it's just brew being not nice.